Move PackagePartProvider to module descriptors.jvm
