+++ compiler/cli/src/org/jetbrains/kotlin/cli/jvm/compiler/CliModuleAnnotationsResolver.kt
@@ -6,8 +6,8 @@
 package org.jetbrains.kotlin.cli.jvm.compiler
 
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
-import org.jetbrains.kotlin.descriptors.PackagePartProvider
 import org.jetbrains.kotlin.js.resolve.getAnnotationsOnContainingJsModule
+import org.jetbrains.kotlin.load.kotlin.PackagePartProvider
 import org.jetbrains.kotlin.load.kotlin.getJvmModuleNameForDeserializedDescriptor
 import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.resolve.ModuleAnnotationsResolver
