+++ core/descriptors.jvm/src/org/jetbrains/kotlin/load/java/lazy/descriptors/LazyJavaPackageFragment.kt
@@ -39,7 +39,7 @@ class LazyJavaPackageFragment(
     private val c = outerContext.childForClassOrPackage(this)
 
     internal val binaryClasses by c.storageManager.createLazyValue {
-        c.components.packageMapper.findPackageParts(fqName.asString()).mapNotNull { partName ->
+        c.components.packagePartProvider.findPackageParts(fqName.asString()).mapNotNull { partName ->
             val classId = ClassId.topLevel(JvmClassName.byInternalName(partName).fqNameForTopLevelClassMaybeWithDollars)
             c.components.kotlinClassFinder.findKotlinClass(classId)?.let { partName to it }
         }.toMap()
